% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{check_folderstruct}
\alias{check_folderstruct}
\title{Check Folder Structure}
\usage{
check_folderstruct()
}
\value{
Throws an error if the folder structure is not correct.
}
\description{
Checks if the current working directory is the correct folder structure for the project.
}
\details{
The current working directory should contain the \code{.Rproj} file and the \code{items} folder.
}
