% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{generate_pw}
\alias{generate_pw}
\title{Passwort Generation}
\usage{
generate_pw(n, len)
}
\arguments{
\item{n}{Number of passworts to generate}

\item{len}{The length of the password}
}
\value{
A character vector
}
\description{
Passwort Generation
}
