% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{generate_users}
\alias{generate_users}
\title{Generate User Credential Data}
\usage{
generate_users(n, len)
}
\arguments{
\item{n}{Numer of users to generate}

\item{len}{Length of the password}
}
\description{
Generate User Credential Data
}
