% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_csv_to_md}
\alias{parse_csv_to_md}
\title{Parse Tidy DataFrame Back to Markdown Without For Loop}
\usage{
parse_csv_to_md(x, item_folder = "items")
}
\arguments{
\item{x}{A data frame containing parsed sections in tidy format.}

\item{item_folder}{folder name, where all items are located (defualts to 'items')}
}
\value{
Writes the reconstructed Markdown to the specified file.
}
\description{
Parse Tidy DataFrame Back to Markdown Without For Loop
}
