% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_md_to_csv}
\alias{parse_md_to_csv}
\title{Parse Markdown to DataFrame}
\usage{
parse_md_to_csv(filename)
}
\arguments{
\item{filename}{The input filename of the Markdown file.}
}
\value{
A data frame containing parsed sections in tidy format.
}
\description{
Parse Markdown to DataFrame
}
