% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push.R
\name{push_all}
\alias{push_all}
\title{Push all items in a directory}
\usage{
push_all(path = "./items", overwrite = FALSE)
}
\arguments{
\item{path}{Path to directory with markdown files (default: "./items")}

\item{overwrite}{Logical, whether to overwrite existing items (default: FALSE)}
}
\description{
Pushes all markdown files in a directory to the database.
If overwrite = TRUE, all existing items will be overwritten without confirmation.
}
