% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{rm_md_comments}
\alias{rm_md_comments}
\title{Remove Markdown Comments in String via Regular Expressions}
\usage{
rm_md_comments(x)
}
\arguments{
\item{x}{A character string.}
}
\value{
A character string without markdown comments.
}
\description{
Remove Markdown Comments in String via Regular Expressions
}
