% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create}
\alias{create}
\title{Create a new Item from the Markdown Template}
\usage{
create(open = TRUE, r_file = NULL)
}
\arguments{
\item{open}{Whether to open the newly created file. Logical vector.}

\item{r_file}{Filename of an R-file (defaults to NULL).}

\item{item_folder}{The folder where the new item should be created. Defaults to "items".}
}
\value{
Creates a new file based on the Markdown-Template with prefilled \code{id_item}.
}
\description{
Create a new Item from the Markdown Template
}
