% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_db}
\alias{update_db}
\title{Update Database}
\usage{
update_db(
  md_file,
  item_folder = "items",
  csv_file = "data_item_tiger.csv",
  sqlite_file = "db_item.sqlite"
)
}
\arguments{
\item{md_file}{path to filename of the newly created item}

\item{item_folder}{folder name, where all items are located (defualts to 'items')}

\item{csv_file}{the csv-file to which the database should be written}

\item{sqlite_file}{the sqlite-file to which the database should be written}
}
\value{
Writes the csv-file and sqlite-file.
}
\description{
Update the sqlite-database and the csv-file containing all items
}
