% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_item}
\alias{validate_item}
\title{Validate a newly created item for correct adherence to formatting}
\usage{
validate_item(filename)
}
\arguments{
\item{filename}{The filename of the md-file of the item}
}
\description{
Validate a newly created item for correct adherence to formatting
}
